lb9/requirements.txt
pandas
scikit-learn
matplotlib
numpy
Pillow
torch
torchvision
torchsummary
grad-cam
seaborn